# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: InfoQuest - Plataforma de Competencia de Conocimientos de Informatica
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user (docente or estudiante)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /challenges/{challengeId}:
    get:
      operationId: getChallenge
      tags: [challenges]
      summary: Get a challenge with its questions
      parameters:
        - name: challengeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Challenge with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeWithQuestions"
  /challenges/custom:
    get:
      operationId: getCustomChallenges
      tags: [challenges]
      summary: List custom challenges created by docentes
      responses:
        "200":
          description: List of custom challenges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomChallenge"
    post:
      operationId: createCustomChallenge
      tags: [challenges]
      summary: Create a new custom challenge (docente only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomChallengeBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomChallenge"
  /challenges/custom/{challengeId}:
    get:
      operationId: getCustomChallenge
      tags: [challenges]
      summary: Get a custom challenge with questions
      parameters:
        - name: challengeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Custom challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeWithQuestions"
    put:
      operationId: updateCustomChallenge
      tags: [challenges]
      summary: Update a custom challenge
      parameters:
        - name: challengeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomChallengeBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomChallenge"
    delete:
      operationId: deleteCustomChallenge
      tags: [challenges]
      summary: Delete a custom challenge
      parameters:
        - name: challengeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    RegisterBody:
      type: object
      properties:
        nombre:
          type: string
        usuario:
          type: string
        password:
          type: string
        rol:
          type: string
          enum: [estudiante, docente]
        grado_bachillerato:
          type: integer
          nullable: true
      required:
        - nombre
        - usuario
        - password
        - rol
    LoginBody:
      type: object
      properties:
        usuario:
          type: string
        password:
          type: string
      required:
        - usuario
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/Profile"
        token:
          type: string
      required:
        - user
        - token
    Profile:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        usuario:
          type: string
        rol:
          type: string
          enum: [estudiante, docente]
        grado_bachillerato:
          type: integer
          nullable: true
        avatar_url:
          type: string
          nullable: true
        puntos_totales:
          type: integer
        retos_completados:
          type: integer
        racha_dias:
          type: integer
        ultimo_acceso:
          type: string
          nullable: true
        creado_en:
          type: string
        ultimo_nivel_intento:
          type: string
          nullable: true
        mejor_puntaje_por_modulo:
          type: object
          nullable: true
      required:
        - id
        - nombre
        - usuario
        - rol
        - puntos_totales
        - retos_completados
        - racha_dias
        - creado_en
    Module:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        anio_bachillerato:
          type: integer
        descripcion:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        icono:
          type: string
          nullable: true
        activo:
          type: boolean
        creado_en:
          type: string
      required:
        - id
        - nombre
        - anio_bachillerato
        - activo
        - creado_en
    Level:
      type: object
      properties:
        id:
          type: integer
        id_modulo:
          type: integer
        nombre:
          type: string
        orden:
          type: integer
        descripcion:
          type: string
          nullable: true
        puntos_por_reto:
          type: integer
        activo:
          type: boolean
      required:
        - id
        - id_modulo
        - nombre
        - orden
        - puntos_por_reto
        - activo
    Question:
      type: object
      properties:
        id:
          type: integer
        id_modulo:
          type: integer
        id_nivel:
          type: integer
        tipo:
          type: string
          enum: [multiple_choice, true_false, code_completion, drag_drop]
        texto:
          type: string
        opciones:
          type: array
          items:
            type: string
          nullable: true
        respuesta_correcta:
          type: string
        explicacion:
          type: string
          nullable: true
        dificultad:
          type: string
          enum: [facil, medio, dificil]
        puntos:
          type: integer
        activa:
          type: boolean
      required:
        - id
        - id_modulo
        - id_nivel
        - tipo
        - texto
        - respuesta_correcta
        - dificultad
        - puntos
        - activa
    Challenge:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        descripcion:
          type: string
          nullable: true
        tipo_juego:
          type: string
          enum: [quiz, code_challenge, security_puzzle, drag_drop, speed_race]
        id_modulo:
          type: integer
        id_nivel:
          type: integer
        tiempo_limite:
          type: integer
        puntos_maximos:
          type: integer
        numero_preguntas:
          type: integer
        activo:
          type: boolean
        creado_en:
          type: string
        modulo:
          $ref: "#/components/schemas/Module"
          nullable: true
        nivel:
          $ref: "#/components/schemas/Level"
          nullable: true
      required:
        - id
        - nombre
        - tipo_juego
        - id_modulo
        - id_nivel
        - tiempo_limite
        - puntos_maximos
        - numero_preguntas
        - activo
        - creado_en
    ChallengeWithQuestions:
      allOf:
        - $ref: "#/components/schemas/Challenge"
        - type: object
          properties:
            preguntas:
              type: array
              items:
                $ref: "#/components/schemas/Question"
          required:
            - preguntas
    CustomChallenge:
      type: object
      properties:
        id:
          type: integer
        id_docente:
          type: integer
        nombre:
          type: string
        descripcion:
          type: string
          nullable: true
        tipo_juego:
          type: string
          enum: [quiz, code_challenge, security_puzzle, drag_drop, speed_race]
        id_modulo:
          type: integer
        id_nivel:
          type: integer
        puntos_maximos:
          type: integer
        numero_preguntas:
          type: integer
        tiempo_limite:
          type: integer
        activo:
          type: boolean
        creado_en:
          type: string
        docente:
          $ref: "#/components/schemas/Profile"
          nullable: true
        modulo:
          $ref: "#/components/schemas/Module"
          nullable: true
        nivel:
          $ref: "#/components/schemas/Level"
          nullable: true
      required:
        - id
        - id_docente
        - nombre
        - tipo_juego
        - id_modulo
        - id_nivel
        - puntos_maximos
        - numero_preguntas
        - tiempo_limite
        - activo
        - creado_en
    CreateCustomChallengeBody:
      type: object
      properties:
        nombre:
          type: string
        descripcion:
          type: string
          nullable: true
        tipo_juego:
          type: string
          enum: [quiz, code_challenge, security_puzzle, drag_drop, speed_race]
        id_modulo:
          type: integer
        id_nivel:
          type: integer
        puntos_maximos:
          type: integer
        numero_preguntas:
          type: integer
        tiempo_limite:
          type: integer
        preguntas:
          type: array
          items:
            $ref: "#/components/schemas/CreateQuestionBody"
      required:
        - nombre
        - tipo_juego
        - id_modulo
        - id_nivel
        - puntos_maximos
        - numero_preguntas
        - tiempo_limite
    CreateQuestionBody:
      type: object
      properties:
        tipo:
          type: string
          enum: [multiple_choice, true_false, code_completion, drag_drop]
        texto:
          type: string
        opciones:
          type: array
          items:
            type: string
          nullable: true
        respuesta_correcta:
          type: string
        explicacion:
          type: string
          nullable: true
        dificultad:
          type: string
          enum: [facil, medio, dificil]
        puntos:
          type: integer
      required:
        - tipo
        - texto
        - respuesta_correcta
        - dificultad
        - puntos
...