# lib/api-zod/src/generated/types/challenge.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InfoQuest - Plataforma de Competencia de Conocimientos de Informatica
 * OpenAPI spec version: 0.1.0
 */
import type { ChallengeTipoJuego } from "./challengeTipoJuego";
import type { Level } from "./level";
import type { Module } from "./module";

export interface Challenge {
  id: number;
  nombre: string;
  descripcion?: string | null;
  tipo_juego: ChallengeTipoJuego;
  id_modulo: number;
  id_nivel: number;
  tiempo_limite: number;
  puntos_maximos: number;
  numero_preguntas: number;
  activo: boolean;
  creado_en: string;
  modulo?: Module | null;
  nivel?: Level | null;
}