# lib/api-zod/src/generated/types/challengeWithQuestions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InfoQuest - Plataforma de Competencia de Conocimientos de Informatica
 * OpenAPI spec version: 0.1.0
 */
import type { Challenge } from "./challenge";
import type { Question } from "./question";

export type ChallengeWithQuestions = Challenge & {
  preguntas: Question[];
};