# lib/api-zod/src/generated/types/createQuestionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InfoQuest - Plataforma de Competencia de Conocimientos de Informatica
 * OpenAPI spec version: 0.1.0
 */
import type { CreateQuestionBodyDificultad } from "./createQuestionBodyDificultad";
import type { CreateQuestionBodyTipo } from "./createQuestionBodyTipo";

export interface CreateQuestionBody {
  tipo: CreateQuestionBodyTipo;
  texto: string;
  opciones?: string[] | null;
  respuesta_correcta: string;
  explicacion?: string | null;
  dificultad: CreateQuestionBodyDificultad;
  puntos: number;
}