# lib/api-zod/src/generated/types/module.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InfoQuest - Plataforma de Competencia de Conocimientos de Informatica
 * OpenAPI spec version: 0.1.0
 */

export interface Module {
  id: number;
  nombre: string;
  anio_bachillerato: number;
  descripcion?: string | null;
  color?: string | null;
  icono?: string | null;
  activo: boolean;
  creado_en: string;
}